Only check if larger range contains the smaller one for Day 4
RangeInclusive implements ExactSizeIterator, which has a .len() method
telling us the size of the range. We can then swap the variables if we
need to so we only check if the smaller range is contained within the
larger one.

However, RangeInclusive is only implemented for 8 and 16 bit integers,
so I had to change my numbers to u16. I think this is prevent the usize
that .len() returns from overflowing on 16 bit machines? Seems a bit
weird imo.