Replace for_each with for loop in Day 5
Trying to do `for i in 0..9` triggered the needless_range_loop lint in
clippy, and using a `for_each` stopped the lint. However, using
`iter_mut` and `enumerate` properly addresses the cause of the lint,
which is that we can access each stack directly, instead of addressing
them with `stacks[i]`.